healthcheck datatypes fixed

+++ src/data/migrations/mysql/db_migration_mysql_v1.0.3.sql
@@ -790,7 +790,7 @@ CREATE INDEX idx_microservice_exec_status_microservice_uuid ON MicroserviceExecS
 ALTER TABLE Fogs ADD COLUMN gps_status VARCHAR(32);
 
 ALTER TABLE Microservices ADD COLUMN cpu_set_cpus TEXT;
-ALTER TABLE Microservices ADD COLUMN memory_limit BIGINT;
+ALTER TABLE Microservices ADD COLUMN memory_limit FLOAT;
 
 CREATE TABLE IF NOT EXISTS MicroserviceHealthChecks (
     id INT AUTO_INCREMENT PRIMARY KEY NOT NULL,
